+++ cmd/livesim2/app/handler_urlgen.go
@@ -191,10 +191,11 @@ type nameWithSelect struct {
 type segmentTimelineType string
 
 const (
-	Number              segmentTimelineType = "nr"
-	TimelineTime        segmentTimelineType = "tlt"
-	TimelineNumber      segmentTimelineType = "tlnr"
-	TimelineTimePattern segmentTimelineType = "tltp"
+	Number                segmentTimelineType = "nr"
+	TimelineTime          segmentTimelineType = "tlt"
+	TimelineNumber        segmentTimelineType = "tlnr"
+	TimelineTimePattern   segmentTimelineType = "tltp"
+	TimelineNumberPattern segmentTimelineType = "tlnrp"
 )
 
 func (s *Server) createInitData(aInfo assetsInfo) (data urlGenData, err error) {
@@ -253,6 +254,9 @@ func createURL(r *http.Request, aInfo assetsInfo, drmCfg *drm.DrmConfig) urlGenD
 	case TimelineTimePattern:
 		data.Stl = TimelineTimePattern
 		sb.WriteString("segtimeline_pattern/")
+	case TimelineNumberPattern:
+		data.Stl = TimelineNumberPattern
+		sb.WriteString("segtimelinenr_pattern/")
 	default:
 		fmt.Printf("Bad stl: %s\n", stl)
 	}